import into Eclipse
Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

generate CUDA code without tinfo

import into Eclipse

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

generate CUDA code without tinfo

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

generate CUDA code without tinfo

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

generate CUDA code without tinfo

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

generate CUDA code without tinfo

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

generate CUDA code without tinfo

import into Eclipse

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

generate CUDA code without tinfo

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

generate CUDA code without tinfo

import into Eclipse

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

set default build command

Setup script copies orcl

OpenCL properties query

need to read properties before tranformation arguments

--amend

redirect props to file

process props from multiple platforms and devices

process props from multiple platforms and devices

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

generate CUDA code without tinfo

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

generate CUDA code without tinfo

import into Eclipse

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

set default build command

Setup script copies orcl

OpenCL properties query

need to read properties before tranformation arguments

--amend

redirect props to file

process props from multiple platforms and devices

process props from multiple platforms and devices

process props from multiple platforms and devices

handle platform properties

remove domain code from opencl transformation

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

generate CUDA code without tinfo

import into Eclipse

Create OpenCL module (initially copy of CUDA module).

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

Setup script copies orcl

set default build command

generate CUDA code without tinfo

import into Eclipse

Change occurrences of cuda to opencl

add OpenCL option to orio_main

add orcl script to launch orio with OpenCL support

set default build command

Setup script copies orcl

OpenCL properties query

need to read properties before tranformation arguments

--amend

redirect props to file

process props from multiple platforms and devices

process props from multiple platforms and devices

process props from multiple platforms and devices

handle platform properties

remove domain code from opencl transformation

remove domain code from opencl transformation

remove L1 cache setting, which is not available in OpenCL

remove L1 cache setting, which is not available in OpenCL

opencl boilerplate and kernel generation

opencl support in ptest_driver

add runtime error checking for OpenCL\

opencl works for matScale

some fixes to reductions

fix types in reduction kernels

add tests

skeleton file for performance measurement with TAU

use build command to build query code

pass tinfo to opencl submodule

only use local memory barrier, not global

add post-command option (for taudb_loadtrial)

metadata

vecHint, sizeHint options

Save OpenCL attributes as TAU metadata

generate unique filenames for profiles

allow filename twice in post-command

OpenCL compile flag arguments were in wrong position

put missing .iteritems() back (I don't know how it got deleted!)

fix transformation argument parsing for platform and device

Insert timer calls (need to fix this to only do this when using TAU)

Move end of execute timer to after blocking copy.

Add vecAXPY example.

Update tau_skeleton.c to use timers around OpenCL executions when OpenCL
is used.

vecMAXPY

WAXPY

ex14FF

Don't output TAU_METADATA except for OpenCL.

Spell 'language' correctly.

ex14FJ

viennacl

+++ orio/main/tuner/ptest_driver.py
@@ -2,9 +2,10 @@
 # To compile and execute the performance-testing code to get the performance cost
 #
 
-import os, time, re
+import os, time, re, datetime, uuid
 
 from orio.main.util.globals import *
+
 
 #-----------------------------------------------------
 
@@ -37,13 +38,13 @@ class PerfTestDriver:
         global perftest_counter 
         perftest_counter += 1 
         
-        if not (language == 'c' or language == 'fortran' or language == 'cuda'):
+        if not (language == 'c' or language == 'fortran' or language == 'cuda' or language == 'opencl'):
             err('orio.main.tuner.ptest_driver: unknown output language specified: %s' % language)
         self.language = language
 
         self.__PTEST_FNAME=Globals().out_prefix+self.__PTEST_FNAME
 
-        if language == 'c': 
+        if language == 'c' or language == 'opencl': 
             self.ext = '.c'
         elif language == 'cuda':
             self.ext = '.cu'
@@ -55,7 +56,7 @@ class PerfTestDriver:
         self.original_src_name = self.__PTEST_FNAME + '_original' + self.ext
         self.exe_name = self.__PTEST_FNAME + '.exe'
         self.original_exe_name = self.__PTEST_FNAME + '_original.exe'
-        if self.language == 'cuda':
+        if self.language == 'cuda' or self.language == 'opencl':
             self.obj_name = self.__PTEST_FNAME + '.o'
             self.original_obj_name = self.__PTEST_FNAME + '_original.o'
 
@@ -106,7 +107,7 @@ class PerfTestDriver:
             err('orio.main.tuner.ptest_driver: cannot open file for writing: %s' % self.src_name)
             
         if self.first:
-            if self.language != 'cuda' and not self.tinfo.timer_file and not (os.path.exists(self.timer_file)):
+            if self.language != 'cuda' and self.language != 'opencl' and not self.tinfo.timer_file and not (os.path.exists(self.timer_file)):
                 # Generate the timing routine file
                 try: 
                     f = open(self.timer_file, 'w')
@@ -214,6 +215,10 @@ class PerfTestDriver:
             if status:
                 err('orio.main.tuner.ptest_driver: failed to compile the test cuda code: "%s"' % cmd)
             cmd = ('%s %s -o %s %s' % (build_cmd, self.extra_compiler_opts, self.exe_name, self.obj_name))
+        elif self.language == 'opencl':
+            cmd = ('%s %s -o %s %s %s' % (build_cmd, self.extra_compiler_opts,
+                                             self.exe_name, self.src_name2, 
+                                             self.tinfo.libs))
         else:
             cmd = ('%s %s -o %s %s %s %s' % (build_cmd, self.extra_compiler_opts,
                                              self.exe_name, self.src_name2, 
@@ -244,6 +249,8 @@ class PerfTestDriver:
 
     def __execute(self):
         '''Execute the test to get the performance costs'''
+
+        Globals().metadata['src_filenames'] = ",".join(Globals().src_filenames)
 
         # check if the executable exists
         if not os.path.exists(self.exe_name):
@@ -294,6 +301,19 @@ class PerfTestDriver:
                 f.close()
             except Exception, e:
                 err('orio.main.tuner.ptest_driver: failed to execute the test code: "%s"\n --> %s: %s' % (cmd,e.__class__.__name__, e))
+                
+            if Globals().post_cmd is not None:
+                try:
+                    cmd = Globals().post_cmd
+                    uniq = "profile-" + datetime.datetime.now().strftime("%y-%m-%d:%H:%S") + '-' + uuid.uuid4().hex
+                    if "%s" in cmd:
+                        cmd = cmd % (uniq, uniq)
+                    status = os.system(cmd) 
+                    if status:
+                        err('orio.main.tuner.ptest_driver: failed to execute the post-command: "%s"' % Globals().post_cmd)
+                except Exception, e:
+                    err('orio.main.tuner.ptest_driver: failed to execute the post-command: "%s"\n --> %s: %s' % (Globals().post_cmd,e.__class__.__name__, e))
+        
                 
             try:
                 if out:
@@ -374,7 +394,6 @@ class PerfTestDriver:
         @param test_code: the code for testing multiple coordinates in the search space
         @return: a dictionary of the times corresponding to each coordinate in the search space
         '''
-
         # write the testing code
         self.__write(test_code, perf_param=perf_param)
